Clean up based on browser/browsing experience
I tested the website on mobile which gave two diferent behaviour on Chrome Beta and normal. This issue was based on the vh of body. I corrected the media query.
I also set the height and role for button to enable uniformity (realtive uniformity, h=3em) and accessibility respectively.
I made the image width and height 490px